upstream: [package-alt] finish localizing serde (#22533)
