+++ external-crates/move/crates/move-package-alt/src/cli/build.rs
@@ -5,7 +5,7 @@
 
 use std::path::PathBuf;
 
-use clap::{Command, Parser, Subcommand};
+use clap::Parser;
 
 use crate::{errors::PackageResult, flavor::Vanilla, package::Package};
