+++ external-crates/move/crates/move-package-alt/src/cli/new.rs
@@ -9,9 +9,8 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use anyhow::{Context, ensure};
-use clap::{Command, Parser, Subcommand};
-use move_core_types::identifier::Identifier;
+use anyhow::Context;
+use clap::Parser;
 use move_package::source_package::layout::SourcePackageLayout;
 
 use crate::{errors::PackageResult, package::PackageName};
