+++ external-crates/move/crates/move-package-alt/src/cli/parse.rs
@@ -8,7 +8,7 @@ use std::path::PathBuf;
 use crate::{
     errors::{FileHandle, PackageResult},
     flavor::Vanilla,
-    package::{lockfile::Lockfiles, manifest::Manifest},
+    package::manifest::Manifest,
     schema::ParsedLockfile,
 };
