+++ external-crates/move/crates/move-package-alt/src/errors/files.rs
@@ -4,7 +4,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use std::{
-    cell::RefCell,
     fmt::{Debug, Display},
     fs, io,
     path::{Path, PathBuf},
@@ -12,8 +11,6 @@ use std::{
 
 use append_only_vec::AppendOnlyVec;
 use codespan_reporting::files::SimpleFile;
-
-use super::PackageResult;
 
 /// A wrapper around [PathBuf] that implements [Display]
 #[derive(Clone)]
