+++ external-crates/move/crates/move-package-alt/src/errors/lockfile_error.rs
@@ -3,7 +3,7 @@
 // Modifications Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{ops::Range, path::PathBuf};
+use std::ops::Range;
 
 use codespan_reporting::{
     diagnostic::{Diagnostic, Label},
@@ -16,7 +16,6 @@ use codespan_reporting::{
 use thiserror::Error;
 
 use super::FileHandle;
-use crate::package::PackageName;
 
 #[derive(Error, Debug)]
 #[error("Invalid lockfile: {kind}")]
