+++ external-crates/move/crates/move-package-alt/src/errors/mod.rs
@@ -7,12 +7,9 @@ mod lockfile_error;
 pub use lockfile_error::LockfileError;
 
 mod located;
-mod thefile;
-pub use located::{Located, Location};
-pub use thefile::TheFile;
+pub use located::Location;
 
 mod files;
-use codespan_reporting::diagnostic::Diagnostic;
 pub use files::{FileHandle, Files};
 use thiserror::Error;
