+++ external-crates/move/crates/move-package-alt/src/flavor/vanilla.rs
@@ -7,21 +7,12 @@
 //! implementation supports no flavor-specific resolvers and stores no
 //! additional metadata in the lockfile.
 
-use std::{
-    collections::{self, BTreeMap},
-    iter::empty,
-    marker::PhantomData,
-    path::{Path, PathBuf},
-};
+use std::iter::empty;
 
 use serde::{Deserialize, Serialize};
 
 use super::MoveFlavor;
-use crate::{
-    dependency::{DependencySet, PinnedDependencyInfo},
-    errors::PackageResult,
-    package::PackageName,
-};
+use crate::dependency::{DependencySet, PinnedDependencyInfo};
 
 /// The [Vanilla] implementation of the [MoveFlavor] trait. This implementation
 /// supports no flavor-specific resolvers and stores no additional metadata in
