+++ external-crates/move/crates/move-package-alt/src/graph/builder.rs
@@ -5,7 +5,6 @@
 
 use std::{
     collections::{BTreeMap, btree_map::Entry},
-    fs::read_to_string,
     path::PathBuf,
     sync::{Arc, Mutex},
 };
