+++ external-crates/move/crates/move-package-alt/src/graph/mod.rs
@@ -3,38 +3,21 @@
 // Modifications Copyright (c) 2026 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{
-    collections::{BTreeMap, btree_map::Entry},
-    fs::read_to_string,
-    path::PathBuf,
-    sync::{Arc, Mutex},
-};
+use std::{collections::BTreeMap, sync::Arc};
 
 use builder::PackageGraphBuilder;
 use derive_where::derive_where;
-use move_core_types::identifier::Identifier;
-use path_clean::PathClean;
-use petgraph::{
-    graph::{DiGraph, NodeIndex},
-    visit::EdgeRef,
-};
-use sha2::{Digest, Sha256};
-use tokio::sync::OnceCell;
-use tracing::{debug, info};
+use petgraph::graph::DiGraph;
+use tracing::debug;
 
 // SPDX-License-Identifier: Apache-2.0
 mod builder;
 mod to_lockfile;
 
 use crate::{
-    dependency::PinnedDependencyInfo,
-    errors::{PackageError, PackageResult},
+    errors::PackageResult,
     flavor::MoveFlavor,
-    package::{
-        EnvironmentName, Package, PackageName, lockfile::Lockfiles, manifest::Manifest,
-        paths::PackagePath,
-    },
-    schema::{LockfileDependencyInfo, PackageID, Pin},
+    package::{EnvironmentName, Package, PackageName, manifest::Manifest, paths::PackagePath},
 };
 
 #[derive(Debug)]
