+++ external-crates/move/crates/move-package-alt/src/package/lockfile.rs
@@ -3,34 +3,15 @@
 // Modifications Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{
-    collections::BTreeMap,
-    ffi::OsString,
-    fmt,
-    fs::read_to_string,
-    path::{Path, PathBuf},
-};
+use std::collections::BTreeMap;
 
-use derive_where::derive_where;
-use move_core_types::identifier::Identifier;
-use serde::{Deserialize, Serialize};
-use serde_spanned::Spanned;
-use toml_edit::{
-    DocumentMut, InlineTable, Item, KeyMut, Table, Value,
-    visit_mut::{VisitMut, visit_table_like_kv_mut, visit_table_mut},
-};
 use tracing::debug;
 
-use super::{
-    EnvironmentName, PackageName,
-    manifest::{Digest, Manifest},
-    paths::PackagePath,
-};
+use super::{EnvironmentName, paths::PackagePath};
 use crate::{
-    dependency::{DependencySet, PinnedDependencyInfo},
-    errors::{FileHandle, LockfileError, PackageError, PackageResult, TheFile},
+    errors::{FileHandle, PackageResult},
     flavor::MoveFlavor,
-    schema::{self, PackageID, ParsedLockfile, Pin, Publication},
+    schema::{PackageID, ParsedLockfile, Pin, Publication},
 };
 
 #[derive(Debug)]
