+++ external-crates/move/crates/move-package-alt/src/test_utils/git.rs
@@ -61,7 +61,6 @@
 //! repository.
 
 use std::{
-    env::consts,
     fs,
     path::{Path, PathBuf},
 };
