+++ external-crates/move/crates/move-package-alt/src/test_utils/mod.rs
@@ -12,9 +12,8 @@ mod paths;
 
 use std::{
     env,
-    ffi::OsStr,
     fmt::Write,
-    fs, os,
+    fs,
     path::{Path, PathBuf},
 };
